requirements.txt
absl-py==0.9.0
amqp==2.6.0
asgiref==3.2.5
astor==0.8.1
backcall==0.1.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
cachetools==4.0.0
celery==4.4.6
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-rest-auth==1.0.4
Django==3.0.4
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-cors-headers==3.5.0
django-rest-multiple-models==2.1.3
djangorestframework==3.11.0
djangorestframework-csv==2.1.0
future==0.18.2
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
importlib-metadata==1.7.0
ipykernel==5.2.0
ipyparallel==6.2.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
jupyter-client==6.1.0
jupyter-core==4.6.3
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kombu==4.6.11
Markdown==3.2.1
matplotlib==3.2.1
metakernel==0.24.3
numpy==1.18.2
oauthlib==3.1.0
oct2py==5.0.4
octave-kernel==0.31.0
opt-einsum==3.2.0
pandas==1.0.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
portalocker==1.5.2
prompt-toolkit==3.0.4
protobuf==3.11.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado==6.0.4
traitlets==4.3.3
unicodecsv==0.14.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.12.1
xlrd==1.2.0
zipp==3.1.0
